% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{datasets}
\alias{datasets}
\alias{elevation}
\alias{habitat}
\alias{stem5}
\alias{stem6}
\alias{taxa}
\alias{tree5}
\alias{tree6}
\alias{vft_4quad}
\alias{tree6_3species}
\title{Datasets from Luquillo, Puerto Rico.}
\description{
These are small datasets for examples, with no value for research. They have
been adapted from the package
\href{https://forestgeo.github.io/fgeo.data/}{fgeo.data}.
}
\examples{
str(elevation)

str(habitat)

str(stem5)

str(stem6)

str(tree5)

str(tree6)

str(tree6_3species)

str(taxa, give.attr = FALSE)

str(vft_4quad)

}
\seealso{
Other datasets: \code{\link{download_data}},
  \code{\link{example_path}}
}
\concept{datasets}
