% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_example.R
\name{path_example}
\alias{path_example}
\title{Path to directory containing example data.}
\usage{
path_example(path = NULL)
}
\arguments{
\item{path}{Path to a file (with extension) from inst/extdata/}
}
\value{
A string giving a path to an example.
}
\description{
Path to directory containing example data.
}
\examples{
path_example()
str(read.csv(path_example("taxa.csv")))

}
\seealso{
Other datasets: \code{\link{datasets}},
  \code{\link{download_data}}, \code{\link{x_example}}
}
\concept{datasets}
