% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{toy_datasets}
\alias{toy_datasets}
\alias{elevation}
\alias{habitat}
\alias{stem5}
\alias{stem6}
\alias{taxa}
\alias{tree5}
\alias{tree6}
\alias{vft_4quad}
\title{Datasets from Luquillo, Puerto Rico.}
\description{
These are small datasets for examples, with no value for research. They have
been adapted from the \strong{fgeo.data} package
(https://forestgeo.github.io/fgeo.data/).
}
\examples{
str(elevation)

str(habitat)

str(vft_4quad)

str(taxa)

str(tree5)

str(tree6)

str(stem5)

str(stem6)
}
