% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_example.R
\name{x_example}
\alias{x_example}
\title{Path to directory containing example data.}
\usage{
x_example(path = NULL)
}
\arguments{
\item{path}{Path to a file (with extension) from inst/extdata/}
}
\value{
A string giving a path to an example.
}
\description{
Path to directory containing example data.
}
\examples{
x_example()
str(read.csv(x_example("taxa.csv")))

}
\seealso{
Other datasets: \code{\link{datasets}},
  \code{\link{download_data}}
}
\concept{datasets}
