% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\title{Download data from fgeo.data.}
\usage{
download_data(.data)
}
\arguments{
\item{.data}{A string giving the name of the dataset to download. The name of
any dataset listed \href{https://github.com/forestgeo/fgeo.data/tree/master/data}{here} (without the
extension).}
}
\value{
A dataset.
}
\description{
Download data from fgeo.data.
}
\examples{
str(download_data("luquillo_stem6_random"))

}
\seealso{
Other datasets: \code{\link{datasets}},
  \code{\link{x_example}}
}
\concept{datasets}
