% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\alias{download_scbi}
\title{Download data from fgeo.data.}
\usage{
download_data(.data)

download_scbi(.data)
}
\arguments{
\item{.data}{A string giving the name of the dataset to download. The name of
any dataset listed at
https://github.com/forestgeo/fgeo.data/tree/master/data (without the
extension).}
}
\value{
A dataset.
}
\description{
Download data from fgeo.data.
}
\examples{
download_data("unique_id")
}
