% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgeo.x-package.R
\docType{package}
\name{fgeo.x-package}
\alias{fgeo.x}
\alias{fgeo.x-package}
\title{fgeo.x: Access Small ForestGEO Datasets For Examples}
\description{
Access small example datasets from Luquillo, a
    ForestGEO site in Puerto Rico
    (<https://forestgeo.si.edu/sites/north-america/luquillo>).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/forestgeo/fgeo.x}
  \item Report bugs at \url{https://github.com/forestgeo/fgeo.x/issues}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{leporem@si.edu} [contractor]

Other contributors:
\itemize{
  \item Jess Zimmerman \email{jesskz@ites.upr.edu} [data contributor]
  \item CTFS-ForestGEO \email{ForestGEO@si.edu} [copyright holder, funder]
}

}
\keyword{internal}
